ChaosTool, bugfixes - added ability to create/update/delete MapTemplates, ReactorTiles, MerchantTemplates, and MonsterTemplates - added an integrity check tab that checks templates for validity - fixed a bug causing an exception to be thrown when a monster attempts to pathfind in an invalid direction - added a LootTable.Mode that allows for selecting between (currently) 2 different loot generation modes

+++ Chaos/Models/Abstractions/IBuyShopSource.cs
@@ -7,7 +7,7 @@ public interface IBuyShopSource : IDialogSourceEntity
     ICollection<Item> ItemsForSale { get; }
     int GetStock(string itemTemplateKey);
     bool HasStock(string itemTemplateKey);
-    void Restock(decimal percent);
+    void Restock(int percent);
 
     bool TryDecrementStock(string itemTemplateKey, int amount);
